room bug.

+++ Project290/Project290/Project290/Games/Solitude/SolitudeObjects/Enemies/TyTaylor.cs
@@ -63,7 +63,7 @@ namespace Project290.Games.Solitude.SolitudeObjects.Enemies
                 }
                 else
                 {
-                    long x = 1 / GameClock.Now * 10000000 * 2000;
+                    float x = (float)(((float)1 / (float)GameClock.Now) * 10000000 * 2000);
                     if (x < 1)
                         x = 1;
                     SolitudeScreen.ship.screen.Score += (uint)(20000 * x);
